+++ packages/monitor/src/Dashboard.ts
@@ -5,24 +5,27 @@ import { fileURLToPath } from 'url'
 import nodeStatic from 'node-static'
 import { Socket } from 'net'
 import process from 'process'
-import { Configuration as ServerConfiguration } from '@hocuspocus/server'
-import moment from 'moment'
 import { Storage } from './Storage'
+import { Collector } from './Collector'
 
 export interface Configuration {
+  collector: Collector | undefined,
+  password: string | undefined,
   path: string,
   port: number | undefined,
   storage: Storage | undefined,
-  serverConfiguration: Partial<ServerConfiguration>,
+  user: string | undefined,
 }
 
 export class Dashboard {
 
   configuration: Configuration = {
+    collector: undefined,
+    password: undefined,
     path: 'dashboard',
     port: undefined,
     storage: undefined,
-    serverConfiguration: {},
+    user: undefined,
   }
 
   websocketServer: WebSocket.Server
@@ -38,6 +41,7 @@ export class Dashboard {
       ...configuration,
     }
 
+    // subscribe to storage additions and send them to the client
     this.configuration.storage?.on('add', data => {
       this.send(JSON.stringify({ data: [data] }))
     })
@@ -73,6 +77,10 @@ export class Dashboard {
     const { path } = this.configuration
 
     if (request.url?.split('/')[1] === path) {
+      if (this.basicAuth(request, response)) {
+        return true
+      }
+
       request.url = request.url.replace(path, '')
 
       const publicPath = join(dirname(fileURLToPath(import.meta.url)), '..', 'dashboard', 'dist')
@@ -104,22 +112,7 @@ export class Dashboard {
     this.connections.set(connection, {})
 
     if (this.configuration.storage) {
-      this.configuration.storage.all().then(data => {
-
-        data.push({
-          key: 'info',
-          timestamp: null,
-          value: {
-            version: process.version,
-            platform: process.platform,
-            started: moment().subtract(process.uptime(), 'second').toISOString(),
-            configuration: this.configuration.serverConfiguration,
-          },
-        })
-
-        setTimeout(() => connection.send(JSON.stringify({ data })), 1000)
-
-      })
+      this.sendInitialDataToClient(connection)
     }
 
     connection.on('close', () => {
@@ -146,5 +139,42 @@ export class Dashboard {
         this.close(connection)
       }
     })
+  }
+
+  private async sendInitialDataToClient(connection: WebSocket) {
+    const data = await this.configuration.storage?.all() || []
+
+    data.push({
+      key: 'info',
+      timestamp: null,
+      value: this.configuration.collector?.info(),
+    })
+
+    setTimeout(() => {
+      connection.send(JSON.stringify({ data }))
+    }, 1000)
+  }
+
+  private basicAuth(request: IncomingMessage, response: ServerResponse) {
+    if (!this.configuration.user || !this.configuration.password) {
+      return false
+    }
+
+    const header = request.headers.authorization || ''
+    const token = header.split(/\s+/).pop() || ''
+    const auth = Buffer.from(token, 'base64').toString()
+    const parts = auth.split(/:/)
+
+    const username = parts.shift()
+    const password = parts.join(':')
+
+    if (this.configuration.user !== username || this.configuration.password !== password) {
+      response.writeHead(401, 'Unauthorized', ['WWW-Authenticate', 'Basic realm="@hocuspocus/monitor"'])
+      response.end()
+
+      return true
+    }
+
+    return false
   }
 }