+++ packages/redis/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-alpha.11](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/redis@1.0.0-alpha.10...@hocuspocus/redis@1.0.0-alpha.11) (2021-03-02)
+
+**Note:** Version bump only for package @hocuspocus/redis
+
+
+
+
+
+# [1.0.0-alpha.10](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/redis@1.0.0-alpha.9...@hocuspocus/redis@1.0.0-alpha.10) (2021-03-02)
+
+**Note:** Version bump only for package @hocuspocus/redis
+
+
+
+
+
+# [1.0.0-alpha.9](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/redis@1.0.0-alpha.8...@hocuspocus/redis@1.0.0-alpha.9) (2021-03-02)
+
+**Note:** Version bump only for package @hocuspocus/redis
+
+
+
+
+
+# [1.0.0-alpha.8](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/redis@1.0.0-alpha.7...@hocuspocus/redis@1.0.0-alpha.8) (2021-03-02)
+
+**Note:** Version bump only for package @hocuspocus/redis
+
+
+
+
+
+# [1.0.0-alpha.7](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/redis@1.0.0-alpha.6...@hocuspocus/redis@1.0.0-alpha.7) (2021-03-02)
+
+**Note:** Version bump only for package @hocuspocus/redis
+
+
+
+
+
 # [1.0.0-alpha.6](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/redis@1.0.0-alpha.5...@hocuspocus/redis@1.0.0-alpha.6) (2021-02-22)
 
 **Note:** Version bump only for package @hocuspocus/redis